requirements.txt
numpy>=1.26
pyyaml>=6.0
pytest>=8.0
gymnasium[toy_text]>=0.29
